+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 7.8.4 (git master)
 * Add support for reading data links (linked fields) configuration from MPP files.
 * Updated to avoid an infinite loop when processing certain corrupt files (Contributed by ninthwaveltd).
+* Update MSPDI generation to ensure MS Project correctly recognises complete tasks without resource assignments.
 
 ## 7.8.3 (24/05/2019)
 * Improve handling of task baseline start, start, baseline finish, finish and slack fields read from FTS files.